backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.11
alembic==1.15.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
pydantic-settings==2.8.1
email-validator==2.2.0
pytest==8.3.5
httpx==0.28.1
